Project density dependent population dynamics

Project density dependent dynamics of a population matrix projection model.

Usage

project_threshold(
  A,
  vector = NULL,
  time = 100,
  return.vec = TRUE,
  threshold = NULL,
  A_DD = NULL
)

Arguments

A

A matrix, or list of matrices.

vector

A numeric vector describiing the age/stage distribution.

time

The number of projection intervals.

return.vec

If TRUE, returns the time series of demographic (st)age vectors.

threshold

The population threshold.

A_DD

A single high density population matrix.

Value

A list with projection results.

Details

A version of popbio::project that will project forward until reaching a threshold population size, at which point it will switch to a specified high-density population matrix.

Author

Chrissy Hernandez chrissy.hernandez@biology.ox.ac.uk

Examples

if (FALSE) { # \dontrun{
# Try out the DD projection function when we use a single matrix:
nohunt_threshold<- project_threshold(pop_mat_100_none, vector=pop_n,
                                    time=100, threshold=1000, A_DD=nohunt_highD)
} # }